src/tf_lite_convertion/requirements.txt
numpy>=1.23,<2.0
tensorflow-cpu>=2.12,<3.0
pandas>=2.0,<3.0